:sparkles: use score instead of rank

+++ model/allow_some_strategy.go
@@ -6,7 +6,7 @@ import (
 
 type AllowSomeStrategy struct {
 	BaseStrategy
-	leastReplicatedRank uint16
+	leastReplicatedRank float32
 	exceedLimit         uint16
 }
